Swagger unix client and updated SDKs

+++ client/python/ned_client/apis/typeclasses_api.py
@@ -80,6 +80,7 @@ class TypeclassesApi(object):
             params[key] = val
         del params['kwargs']
 
+
         resource_path = '/typeclasses'.replace('{format}', 'json')
         method = 'GET'
 
@@ -157,6 +158,7 @@ class TypeclassesApi(object):
             params[key] = val
         del params['kwargs']
 
+
         resource_path = '/typeclasses'.replace('{format}', 'json')
         method = 'POST'
 
@@ -218,9 +220,6 @@ class TypeclassesApi(object):
                  If the method is called asynchronously,
                  returns the request thread.
         """
-        # verify the required parameter 'id' is set
-        if id is None:
-            raise ValueError("Missing the required parameter `id` when calling `read`")
 
         all_params = ['id']
         all_params.append('callback')
@@ -234,6 +233,10 @@ class TypeclassesApi(object):
                 )
             params[key] = val
         del params['kwargs']
+
+        # verify the required parameter 'id' is set
+        if ('id' not in params) or (params['id'] is None):
+            raise ValueError("Missing the required parameter `id` when calling `read`")
 
         resource_path = '/typeclasses/{id}'.replace('{format}', 'json')
         method = 'GET'
@@ -310,6 +313,7 @@ class TypeclassesApi(object):
             params[key] = val
         del params['kwargs']
 
+
         resource_path = '/typeclasses/{id}'.replace('{format}', 'json')
         method = 'PUT'
 
@@ -371,9 +375,6 @@ class TypeclassesApi(object):
                  If the method is called asynchronously,
                  returns the request thread.
         """
-        # verify the required parameter 'id' is set
-        if id is None:
-            raise ValueError("Missing the required parameter `id` when calling `delete`")
 
         all_params = ['id']
         all_params.append('callback')
@@ -387,6 +388,10 @@ class TypeclassesApi(object):
                 )
             params[key] = val
         del params['kwargs']
+
+        # verify the required parameter 'id' is set
+        if ('id' not in params) or (params['id'] is None):
+            raise ValueError("Missing the required parameter `id` when calling `delete`")
 
         resource_path = '/typeclasses/{id}'.replace('{format}', 'json')
         method = 'DELETE'
@@ -449,9 +454,6 @@ class TypeclassesApi(object):
                  If the method is called asynchronously,
                  returns the request thread.
         """
-        # verify the required parameter 'typeclassid' is set
-        if typeclassid is None:
-            raise ValueError("Missing the required parameter `typeclassid` when calling `get_global_type_for_type_class`")
 
         all_params = ['typeclassid']
         all_params.append('callback')
@@ -465,6 +467,10 @@ class TypeclassesApi(object):
                 )
             params[key] = val
         del params['kwargs']
+
+        # verify the required parameter 'typeclassid' is set
+        if ('typeclassid' not in params) or (params['typeclassid'] is None):
+            raise ValueError("Missing the required parameter `typeclassid` when calling `get_global_type_for_type_class`")
 
         resource_path = '/typeclasses/{typeclassid}/typevalues'.replace('{format}', 'json')
         method = 'GET'
@@ -528,9 +534,6 @@ class TypeclassesApi(object):
                  If the method is called asynchronously,
                  returns the request thread.
         """
-        # verify the required parameter 'typeclassid' is set
-        if typeclassid is None:
-            raise ValueError("Missing the required parameter `typeclassid` when calling `create_global_type`")
 
         all_params = ['typeclassid', 'body']
         all_params.append('callback')
@@ -544,6 +547,10 @@ class TypeclassesApi(object):
                 )
             params[key] = val
         del params['kwargs']
+
+        # verify the required parameter 'typeclassid' is set
+        if ('typeclassid' not in params) or (params['typeclassid'] is None):
+            raise ValueError("Missing the required parameter `typeclassid` when calling `create_global_type`")
 
         resource_path = '/typeclasses/{typeclassid}/typevalues'.replace('{format}', 'json')
         method = 'POST'
@@ -609,12 +616,6 @@ class TypeclassesApi(object):
                  If the method is called asynchronously,
                  returns the request thread.
         """
-        # verify the required parameter 'typeclassid' is set
-        if typeclassid is None:
-            raise ValueError("Missing the required parameter `typeclassid` when calling `get_global_type`")
-        # verify the required parameter 'typevalueid' is set
-        if typevalueid is None:
-            raise ValueError("Missing the required parameter `typevalueid` when calling `get_global_type`")
 
         all_params = ['typeclassid', 'typevalueid']
         all_params.append('callback')
@@ -628,6 +629,13 @@ class TypeclassesApi(object):
                 )
             params[key] = val
         del params['kwargs']
+
+        # verify the required parameter 'typeclassid' is set
+        if ('typeclassid' not in params) or (params['typeclassid'] is None):
+            raise ValueError("Missing the required parameter `typeclassid` when calling `get_global_type`")
+        # verify the required parameter 'typevalueid' is set
+        if ('typevalueid' not in params) or (params['typevalueid'] is None):
+            raise ValueError("Missing the required parameter `typevalueid` when calling `get_global_type`")
 
         resource_path = '/typeclasses/{typeclassid}/typevalues/{typevalueid}'.replace('{format}', 'json')
         method = 'GET'
@@ -694,12 +702,6 @@ class TypeclassesApi(object):
                  If the method is called asynchronously,
                  returns the request thread.
         """
-        # verify the required parameter 'typeclassid' is set
-        if typeclassid is None:
-            raise ValueError("Missing the required parameter `typeclassid` when calling `update_global_type`")
-        # verify the required parameter 'typevalueid' is set
-        if typevalueid is None:
-            raise ValueError("Missing the required parameter `typevalueid` when calling `update_global_type`")
 
         all_params = ['typeclassid', 'typevalueid', 'body']
         all_params.append('callback')
@@ -713,6 +715,13 @@ class TypeclassesApi(object):
                 )
             params[key] = val
         del params['kwargs']
+
+        # verify the required parameter 'typeclassid' is set
+        if ('typeclassid' not in params) or (params['typeclassid'] is None):
+            raise ValueError("Missing the required parameter `typeclassid` when calling `update_global_type`")
+        # verify the required parameter 'typevalueid' is set
+        if ('typevalueid' not in params) or (params['typevalueid'] is None):
+            raise ValueError("Missing the required parameter `typevalueid` when calling `update_global_type`")
 
         resource_path = '/typeclasses/{typeclassid}/typevalues/{typevalueid}'.replace('{format}', 'json')
         method = 'PUT'
@@ -780,12 +789,6 @@ class TypeclassesApi(object):
                  If the method is called asynchronously,
                  returns the request thread.
         """
-        # verify the required parameter 'typeclassid' is set
-        if typeclassid is None:
-            raise ValueError("Missing the required parameter `typeclassid` when calling `delete_global_type`")
-        # verify the required parameter 'typevalueid' is set
-        if typevalueid is None:
-            raise ValueError("Missing the required parameter `typevalueid` when calling `delete_global_type`")
 
         all_params = ['typeclassid', 'typevalueid']
         all_params.append('callback')
@@ -799,6 +802,13 @@ class TypeclassesApi(object):
                 )
             params[key] = val
         del params['kwargs']
+
+        # verify the required parameter 'typeclassid' is set
+        if ('typeclassid' not in params) or (params['typeclassid'] is None):
+            raise ValueError("Missing the required parameter `typeclassid` when calling `delete_global_type`")
+        # verify the required parameter 'typevalueid' is set
+        if ('typevalueid' not in params) or (params['typevalueid'] is None):
+            raise ValueError("Missing the required parameter `typevalueid` when calling `delete_global_type`")
 
         resource_path = '/typeclasses/{typeclassid}/typevalues/{typevalueid}'.replace('{format}', 'json')
         method = 'DELETE'
